+++ lib/features/chat/domain/entities/chat_entity.dart
