+++ lib/features/chat/domain/entities/user_entity.dart
