+++ lib/features/chat/domain/repositories/user_repository.dart
@@ -1,0 +1,16 @@
+import 'package:flutter_chat/features/chat/domain/entities/chat_entity.dart';
+import 'package:flutter_chat/features/chat/domain/entities/message_entity.dart';
+import 'package:flutter_chat/features/chat/domain/entities/user_entity.dart';
+
+abstract class UserRepository {
+  Stream<List<UserEntity>> users(String username);
+  Future<UserEntity> getUser(String username);
+  Future addMessage(
+      String chatRoomId, String messageId, Map<String, dynamic> messageInfoMap);
+  Future<void> updateLastMessageSend(
+      String chatRoomId, Map<String, dynamic> lastMessageInfoMap);
+  Future<void> createChatRoom(
+      String chatRoomId, Map<String, dynamic> chatRoomInfoMap);
+  Stream<List<MessageEntity>> getChatRoomMessages(String chatRoomId);
+  Stream<List<ChatEntity>> getChatRooms(String myUsername);
+}
