--- lib/features/search_user_for_chat/presentation/bloc/user_bloc/user_state.dart
@@ -1,25 +1,0 @@
-import 'package:equatable/equatable.dart';
-import 'package:flutter_chat/features/search_user_for_chat/domain/entities/user_entity.dart';
-
-abstract class UsersState extends Equatable {
-  const UsersState();
-
-  @override
-  List<Object> get props => [];
-}
-
-class UsersLoadInProgress extends UsersState {}
-
-class UsersLoadSuccess extends UsersState {
-  final List<UserEntity> users;
-
-  const UsersLoadSuccess({required this.users});
-
-  @override
-  List<Object> get props => [users];
-
-  @override
-  String toString() => 'UsersLoadSuccess { todos: $users }';
-}
-
-class UsersLoadFailure extends UsersState {}
