+++ lib/features/sign_in/presentation/bloc/sign_in_with_google/sign_in_with_google_event.dart
@@ -1,0 +1,13 @@
+import 'package:equatable/equatable.dart';
+
+abstract class SignInWithGoogleEvent extends Equatable {}
+
+class AppStarted extends SignInWithGoogleEvent {
+  @override
+  List<Object> get props => [];
+}
+
+class AppExit extends SignInWithGoogleEvent {
+  @override
+  List<Object> get props => [];
+}
