--- lib/features/sign_in_with_google/presentation/bloc/sign_in_with_google_bloc/sign_in_with_google_event.dart
@@ -1,13 +1,0 @@
-import 'package:equatable/equatable.dart';
-
-abstract class AuthenticationEvent extends Equatable {}
-
-class AppStarted extends AuthenticationEvent {
-  @override
-  List<Object> get props => [];
-}
-
-class AppExit extends AuthenticationEvent {
-  @override
-  List<Object> get props => [];
-}
