--- lib/features/sign_in_with_google/presentation/bloc/sign_in_with_google_bloc/sign_in_with_google_state.dart
@@ -1,28 +1,0 @@
-import 'package:equatable/equatable.dart';
-
-abstract class AuthenticationState extends Equatable {
-  const AuthenticationState();
-
-  @override
-  List<Object> get props => [];
-}
-
-class Uninitialized extends AuthenticationState {}
-
-class Authenticated extends AuthenticationState {
-  const Authenticated(this.userId);
-
-  final String userId;
-
-  @override
-  List<Object> get props => [userId];
-
-  @override
-  String toString() => 'Authenticated { userId: $userId }';
-}
-
-class Unauthenticated extends AuthenticationState {}
-
-class SignOuted extends AuthenticationState {}
-
-class UnSignOut extends AuthenticationState {}
