+++ lib/locator_service.dart
@@ -1,12 +1,13 @@
-import 'package:flutter_chat/features/search_user_for_chat/data/repositories/user_repository_impl.dart';
-import 'package:flutter_chat/features/search_user_for_chat/domain/repositories/user_repository.dart';
-import 'package:flutter_chat/features/search_user_for_chat/presentation/bloc/user_bloc/user_bloc.dart';
-import 'package:flutter_chat/features/search_user_for_chat/presentation/bloc/user_info_bloc/user_info_bloc.dart';
-import 'package:flutter_chat/features/sign_in_with_google/data/repositories/user_signin_repository_impl.dart';
-import 'package:flutter_chat/features/sign_in_with_google/domain/repositories/user_signin_repository.dart';
-import 'package:flutter_chat/features/sign_in_with_google/presentation/bloc/sign_in_with_google_bloc/sign_in_with_google_bloc.dart';
+import 'package:flutter_chat/features/chat/data/repositories/user_repository_impl.dart';
+import 'package:flutter_chat/features/chat/domain/repositories/user_repository.dart';
+import 'package:flutter_chat/features/chat/presentation/bloc/user_info_bloc/user_info_bloc.dart';
+import 'package:flutter_chat/features/sign_in/data/repositories/user_signin_repository_impl.dart';
+import 'package:flutter_chat/features/sign_in/domain/repositories/user_signin_repository.dart';
+import 'package:flutter_chat/features/sign_in/presentation/bloc/sign_in_with_google/sign_in_with_google_bloc.dart';
 import 'package:get_it/get_it.dart';
 import 'package:shared_preferences/shared_preferences.dart';
+
+import 'features/chat/presentation/bloc/user_bloc/user_bloc.dart';
 
 final sl = GetIt.instance;
