+++ lib/main.dart
@@ -2,12 +2,12 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_chat/features/sign_in_with_google/presentation/bloc/sign_in_with_google_bloc/sign_in_with_google_bloc.dart';
-import 'package:flutter_chat/features/sign_in_with_google/presentation/bloc/sign_in_with_google_bloc/sign_in_with_google_event.dart';
+import 'package:flutter_chat/features/chat/presentation/bloc/user_info_bloc/user_info_bloc.dart';
+import 'package:flutter_chat/features/sign_in/presentation/bloc/sign_in_with_google/sign_in_with_google_event.dart';
 import 'package:flutter_chat/locator_service.dart' as di;
 import 'package:flutter_chat/main_application_screen.dart';
-import 'features/search_user_for_chat/presentation/bloc/user_bloc/user_bloc.dart';
-import 'features/search_user_for_chat/presentation/bloc/user_info_bloc/user_info_bloc.dart';
+import 'features/chat/presentation/bloc/user_bloc/user_bloc.dart';
+import 'features/sign_in/presentation/bloc/sign_in_with_google/sign_in_with_google_bloc.dart';
 
 void main() async {
   SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(
